:truck: move ptt crawl related packages in ptt package
