+++ ptt/web/crawler_test.go
@@ -1,4 +1,4 @@
-package crawler
+package web
 
 import (
 	"reflect"
